Add files via upload
Python files that unzips gzip JSON files and extract data from the records to format selected data as the output in a Redis database. So far, it is believed that the functionality of  get_WatchEvent(), get_ForkEvent(), get_ReleaseEvents(), get_MemberEvent(), get_IssuesEvent() works fine. Although a modification for get_ForkEvent() could be added to account for the case in which the record format uses the key, 'repository' as opposed to 'repo'. Currently, when running the method, get_PushEvent(), an error that says, 'MISCONF Redis is configured to save RDB snapshots, but it is currently not able to persist on disk...' occurs. In this commit, the code in repos.py and person.py still needs to be updated to use the Redis functionality as opposed to the locking mechanism.